requirements.txt
diffusers==0.24.0
numpy==1.26.2
openai==1.5.0
Pillow==10.1.0
torch==2.1.2
triton==2.1.0
xformers==0.0.23.post1
